+++ app/page.tsx
@@ -1,124 +1,16 @@
 "use client";
 import Hero from "@common/hero/Hero";
 import "./globals.css";
-import { HTMLInputTypeAttribute, ReactNode, useState } from "react";
+import { useState } from "react";
 import { executeAlchemyApiWithParams } from "@common/utils/alchemy";
-import { TransactionParams, mockParams } from "@common/utils/mocks";
-import Button from "@common/Button";
-import formatWalletAddress from "@common/utils/formatWalletAddress";
-
-const MockupCode = ({ children }: { children: ReactNode }) => {
-  return (
-    <div className="mockup-code w-full h-full rounded-xl overflow-auto">
-      {children}
-    </div>
-  )
-}
-
-const MockupCodeComponent = ({ children, center = false }: { children: ReactNode, center?: boolean }) => {
-  if (center) {
-    return (
-      <MockupCode>
-        <div className="w-full h-full items-center justify-center flex flex-col">
-          {children}
-        </div>
-      </MockupCode>
-    )
-  }
-  return (
-    <MockupCode>
-      <pre>
-        {children}
-      </pre>
-    </MockupCode>
-  )
-}
-
-const DataDisplay = ({text, loading}: {text: string | null, loading?: boolean}) => {
-  if (loading) {
-    return (
-      <MockupCodeComponent center>
-        <span className="loading loading-ring loading-lg"></span>
-      </MockupCodeComponent>
-    )
-  }
-  return (
-    <MockupCodeComponent>
-      {text || "Simulate a transaction to begin"}
-    </MockupCodeComponent>
-  )
-}
-const InputTypeSelector = ({
-  text,
-  value,
-  onChecked,
-  checked,
-  name,
-  type,
-  styles
-}: {
-  text: string,
-  value: string | boolean,
-  onChecked: (ch: any) => void
-  checked: boolean
-  name: string
-  type: HTMLInputTypeAttribute
-  styles: string
-}) => {
-  return (
-    <div className="form-control">
-      <label className="flex flex-row gap-3 label cursor-pointer">
-        <span className="label-text">
-          {text}
-        </span>
-        <input
-          type={type}
-          name={name}
-          className={styles}
-          onChange={type === "radio" ? (change) => onChecked(value): (change) => onChecked(change.target.checked)}
-          checked={checked}
-        />
-      </label>
-    </div>
-  )
-}
+import { mockParams } from "@common/utils/mocks";
+import Button from "@common/components/Button";
+import { DataDisplay } from "@common/components/MockupCode";
+import { InputTypeSelector } from "@common/components/InputTypeSelector";
+import { TransactionSelector } from "@common/components/TransactionSelector";
 
 const DEFUALT_PARAMS = mockParams.simulateAssetChanges[0];
-type ExecutionType = "SIMULATE_EXECUTION"|"SIMULATE_ASSET_CHANGES";
-
-const TransactionSelector = ({
-  setParams,
-  transactions
-}: {
-  setParams: (p: any) => void
-  transactions: TransactionParams
-}) => {
-  return (
-    <>
-    {
-      transactions.map((tx, key) => {
-        return (
-          <div
-            className="card w-full bg-base-200 border-2 border-primary-focus mb-3"
-            key={key}
-          >
-            <div className="card-body text-sm">
-              <p>{tx.method}</p>
-              <p>
-                From: {formatWalletAddress(tx.params[0].from)}
-              </p>
-              <p>
-                To: {formatWalletAddress(tx.params[0].to)}
-              </p>
-            </div>
-          </div>
-        )
-      })
-    }
-    </>
-  )
-}
-
+type ExecutionType = "SIMULATE_EXECUTION" | "SIMULATE_ASSET_CHANGES";
 export default function Home() {
   const [executionType, setExecutionType] = useState<ExecutionType>("SIMULATE_EXECUTION");
